requirements-controlnet.txt
numpy
pillow
torch>=2.4
torchvision>=0.19
transformers>=4.44
diffusers
accelerate
safetensors
opencv-python
